finished the searching for request modal, deleted Sanjana's version of it
